+++ src/mailman/docs/NEWS.rst
@@ -36,6 +36,10 @@ Interfaces
    Given by Aurélien Bompard, tweaked by Barry Warsaw.
  * The default `postauth.txt` and `postheld.txt` templates now no longer
    include the inaccurate admindb and confirmation urls.
+ * Messages now include a `Message-ID-Hash` as the replacement for
+   `X-Message-ID-Hash` although the latter is still included for backward
+   compatibility.  Also be sure that all places which add the header use the
+   same algorithm.
 
 Internal API
 ------------